Move shakespeare to benchmark folder.
PiperOrigin-RevId: 304729158
