+++ official/benchmark/shakespeare_benchmark.py
@@ -23,13 +23,13 @@ import time
 from absl import flags
 import tensorflow as tf  # pylint: disable=g-bad-import-order
 
-from official.staging.shakespeare import shakespeare_main
+from official.benchmark.models.shakespeare import shakespeare_main
 from official.utils.flags import core as flags_core
 from official.utils.misc import keras_utils
 from official.utils.testing import benchmark_wrappers
 from official.utils.testing.perfzero_benchmark import PerfZeroBenchmark
 
-SHAKESPEARE_TRAIN_DATA = 'shakespeare/shakespeare.txt'
+SHAKESPEARE_TRAIN_DATA = '/placer/prod/home/tensorflow-performance-data/datasets/shakespeare/shakespeare.txt'
 TMP_DIR = os.getenv('TMPDIR')
 FLAGS = flags.FLAGS
